[feat] 모듈 분리
