+++ app/src/main/java/woowacourse/shopping/DIApplication.kt
@@ -1,15 +1,9 @@
 package woowacourse.shopping
 
 import android.app.Application
-import woowacourse.shopping.data.repository.CartRepositoryImpl
-import woowacourse.shopping.data.repository.ProductRepositoryImpl
+import com.angrypig.autodi.AutoDI
 import woowacourse.shopping.di.repositoryModule
 import woowacourse.shopping.di.viewModelModule
-import woowacourse.shopping.repository.CartRepository
-import woowacourse.shopping.repository.ProductRepository
-import woowacourse.shopping.ui.MainViewModel
-import woowacourse.shopping.util.autoDI.AutoDI
-import woowacourse.shopping.util.autoDI.autoDIModule.autoDIModule
 
 class DIApplication : Application() {
     override fun onCreate() {
