+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -3,10 +3,9 @@ package woowacourse.shopping.ui.cart
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
-import woowacourse.shopping.util.autoDI.injectViewModel
+import com.angrypig.autodi.injectViewModel
 
 class CartActivity : AppCompatActivity() {
