--- app/src/main/java/woowacourse/shopping/util/autoDI/Injector.kt
@@ -1,10 +1,0 @@
-package woowacourse.shopping.util.autoDI
-
-import kotlin.reflect.KProperty
-
-class Injector(val qualifier: String? = null) {
-    internal inline operator fun <reified T : Any> getValue(thisRef: Any?, property: KProperty<*>): T {
-        @Suppress("UNCHECKED_CAST")
-        return AutoDI.inject(qualifier)
-    }
-}
