--- app/src/main/java/woowacourse/shopping/util/autoDI/Reflects.kt
@@ -1,10 +1,0 @@
-package woowacourse.shopping.util.autoDI
-
-import kotlin.reflect.KFunction
-import kotlin.reflect.KType
-import kotlin.reflect.jvm.reflect
-
-internal fun <T> getLambdaReturnType(lambda: () -> T): KType? {
-    val reflectedLambda = lambda.reflect() ?: return null
-    return (reflectedLambda as? KFunction<*>)?.returnType
-}
