--- app/src/main/java/woowacourse/shopping/util/autoDI/autoDIModule/AutoDiModules.kt
@@ -1,7 +1,0 @@
-package woowacourse.shopping.util.autoDI.autoDIModule
-
-fun autoDIModule(qualifier: String? = null, init: AutoDIModule.() -> Unit): AutoDIModule {
-    val module = AutoDIModule(qualifier)
-    module.init()
-    return module
-}
