--- app/src/main/java/woowacourse/shopping/util/autoDI/autoDIModule/register/DisposableRegister.kt
@@ -1,9 +1,0 @@
-package woowacourse.shopping.util.autoDI.autoDIModule.register
-
-import woowacourse.shopping.util.autoDI.autoDIModule.LifeCycleTypes
-
-class DisposableRegister(private val disposables: LifeCycleTypes.Disposables) {
-    internal fun <T : Any> register(qualifier: String? = null, initializeMethod: () -> T) {
-        disposables.add(qualifier, initializeMethod)
-    }
-}
