+++ autoDI/src/main/java/com/angrypig/autodi/autoDIModule/AutoDiModules.kt
@@ -1,0 +1,7 @@
+package com.angrypig.autodi.autoDIModule
+
+fun autoDIModule(qualifier: String? = null, init: AutoDIModule.() -> Unit): AutoDIModule {
+    val module = AutoDIModule(qualifier)
+    module.init()
+    return module
+}
