+++ autoDI/src/main/java/com/angrypig/autodi/autoDIModule/register/DisposableRegister.kt
@@ -1,0 +1,9 @@
+package com.angrypig.autodi.autoDIModule.register
+
+import com.angrypig.autodi.autoDIModule.LifeCycleTypes
+
+class DisposableRegister(private val disposables: LifeCycleTypes.Disposables) {
+    internal fun <T : Any> register(qualifier: String? = null, initializeMethod: () -> T) {
+        disposables.add(qualifier, initializeMethod)
+    }
+}
