+++ settings.gradle.kts
@@ -15,3 +15,4 @@ dependencyResolutionManagement {
 rootProject.name = "android-di"
 include(":app")
 include(":domain")
+include(":autoDI")
